services/shared-python/requirements.txt
# Arasul Shared Python Library Dependencies

# Database
psycopg2-binary>=2.9.9

# HTTP Client
requests>=2.31.0
urllib3>=2.0.0

# Health Check Server
flask>=3.0.0

# Async Support (optional)
aiohttp>=3.9.0
